backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
httpx==0.27.0
tavily-python==0.5.0
openai>=1.60.0
neo4j==5.25.0
python-dotenv==1.0.1
pydantic==2.9.0
sse-starlette==2.1.0
asyncpg==0.29.0
